Support cluster id (#687)
We use cluster id to verify whether one host belongs to our cluster or not. 
If the host send heart beat to metad without clusterId,  metad will take it as a new host into our cluster.  Otherwise, metad will check the cluster id received is equals to the local one,  and reject the host request if cluster id conflicts. 

After the cluster id supported,  we don't really need add/remove hosts steps anymore.

+++ src/interface/raftex.thrift
@@ -35,7 +35,7 @@ enum ErrorCode {
     E_EXCEPTION = -20;          // An thrift internal exception was thrown
 }
 
-typedef byte (cpp.type = "nebula::ClusterID") ClusterID
+typedef i64 (cpp.type = "nebula::ClusterID") ClusterID
 typedef i32 (cpp.type = "nebula::GraphSpaceID") GraphSpaceID
 typedef i32 (cpp.type = "nebula::PartitionID") PartitionID
 typedef i64 (cpp.type = "nebula::TermID") TermID
